requirements.txt
pandas
watchdog
mysql-connector-python